Update README.md
This may not be the best way to provide clarity here, but as a 1st time user of the guide the intended placement location of the index.page file was unclear, and my 1st run produced errors. I understand this is alpha, but figure I'd comment anyway in case this 1st experience is useful.